Refactor: #71 SubView 분리
