+++ NoMyeolmang/NoMyeolmang/Sources/Presentation/ReportView/ReportView.swift
@@ -13,8 +13,6 @@ struct ReportView: View {
     @ObservedObject var viewModel: ReportViewModel
 
     var body: some View {
-        let boxSize = CGSize(width: 345, height: 119)
-
         ZStack {
             Image("backgroundSpace")
                 .resizable()
@@ -36,49 +34,7 @@ struct ReportView: View {
                     .foregroundColor(.white)
                     .padding(.top,1)
 
-                HStack(spacing: 48){
-                    VStack(spacing: 2){
-                        Text("지구 시간")
-                            .textStyle(GSFont.Light14)
-                            .foregroundColor(.white)
-                        Text(viewModel.earthTime)
-                            .textStyle(GSFont.SemiBold28)
-                            .foregroundColor(.white)
-                        }
-                    Rectangle()
-                        .fill(Color.white)
-                        .frame(width: 1, height: 99)
-                        .padding(.top, 10)
-                        .padding(.bottom, 10)
-                        .opacity(0.1)
-                    VStack(spacing: 2){
-                        Text("탐사 시간")
-                            .textStyle(GSFont.Light14)
-                            .foregroundColor(.white)
-                        Text(viewModel.focusTime)
-                            .textStyle(GSFont.SemiBold28)
-                            .foregroundColor(.white)
-                        }
-                    } // hstack
-                
-                .frame(width: boxSize.width, height: boxSize.height, alignment: .top)
-                .background(
-                    LinearGradient(
-                        gradient: Gradient(colors: [
-                            Color("939393").opacity(0.02),
-                            Color("A471C8").opacity(0.3)
-                        ]),
-                        startPoint: .top,
-                        endPoint: .bottom
-                    )
-                    .opacity(0.4)
-                    .clipShape(RoundedRectangle(cornerRadius: 10))
-                )
-                .overlay(
-                    RoundedRectangle(cornerRadius: 10)
-                        .stroke(Color.white.opacity(0.7), lineWidth: 1)
-                )
-                .padding(.top, 25)
+                TimerReportBox(earthTime: viewModel.earthTime, focusTime: viewModel.focusTime)
                 
                 
                 GSButton(title: "다시 시작하기", width: 250) {
